+++ context/app/static/js/components/Providers.jsx
@@ -8,7 +8,6 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import GlobalStyles from 'js/components/globalStyles';
 import theme from '../theme';
 import GlobalFonts from '../fonts';
-import MatomoProviderWrapper from './MatomoProviderWrapper';
 
 const generateClassName = createGenerateClassName({
   disableGlobal: true,
@@ -21,20 +20,18 @@ function Providers(props) {
   const { endpoints, groupsToken, children } = props;
   // injectFirst ensures styled-components takes priority over mui for styling
   return (
-    <MatomoProviderWrapper>
-      <StylesProvider generateClassName={generateClassName} injectFirst>
-        <GlobalFonts />
-        <MuiThemeProvider theme={theme}>
-          <ThemeProvider theme={theme}>
-            <AppContext.Provider value={{ groupsToken, ...endpoints }}>
-              <CssBaseline />
-              <GlobalStyles />
-              {children}
-            </AppContext.Provider>
-          </ThemeProvider>
-        </MuiThemeProvider>
-      </StylesProvider>
-    </MatomoProviderWrapper>
+    <StylesProvider generateClassName={generateClassName} injectFirst>
+      <GlobalFonts />
+      <MuiThemeProvider theme={theme}>
+        <ThemeProvider theme={theme}>
+          <AppContext.Provider value={{ groupsToken, ...endpoints }}>
+            <CssBaseline />
+            <GlobalStyles />
+            {children}
+          </AppContext.Provider>
+        </ThemeProvider>
+      </MuiThemeProvider>
+    </StylesProvider>
   );
 }
