+++ context/app/static/js/components/Routes/useSendPageView.js
@@ -1,37 +1,29 @@
 import { useEffect } from 'react';
-import ReactGA from 'react-ga';
-import { useMatomo } from '@datapunt/matomo-tracker-react';
+import { trackPageView } from 'js/helpers/trackers';
 
 function useSendPageView(path) {
-  const { trackPageView } = useMatomo();
-
   useEffect(() => {
-    function track(trackPath) {
-      ReactGA.pageview(trackPath);
-      trackPageView({ href: trackPath });
-    }
-
     if (path.startsWith('/browse')) {
       const trackPath = path.match(/\/browse\/\w+/)[0];
-      track(trackPath);
+      trackPageView(trackPath);
       return;
     }
     // send path without ID for specific saved list page
     if (path.startsWith('/my-lists/')) {
       // Distinguished by final slash.
       const trackPath = '/my-lists/saved-list';
-      track(trackPath);
+      trackPageView(trackPath);
       return;
     }
     if (path.startsWith('/search') || path.startsWith('/dev-search')) {
       const urlParams = new URLSearchParams(window.location.search);
       const entityTypeKey = 'entity_type[0]';
       const trackPath = `${path}?${entityTypeKey}=${urlParams.get(entityTypeKey)}`;
-      track(trackPath);
+      trackPageView(trackPath);
       return;
     }
-    track(path);
-  }, [path, trackPageView]);
+    trackPageView(path);
+  }, [path]);
 }
 
 export default useSendPageView;
